Changed 'dotsPerPixel' from int to float. Reverted BreakIterator from a word instance back to line instance.

+++ flying-saucer-core/src/main/java/org/xhtmlrenderer/css/style/CssContext.java
@@ -15,7 +15,7 @@ import org.xhtmlrenderer.render.FSFontMetrics;
 public interface CssContext {
     float getMmPerDot();
     
-    int getDotsPerPixel();
+    float getDotsPerPixel();
 
     float getFontSize2D(FontSpecification font);
